refactor(frontend): uploader to upload to s3
